+++ cmd/jaeger/internal/extension/jaegerquery/internal/mcptools/internal/handlers/search_traces.go
@@ -181,7 +181,7 @@ func (h *searchTracesHandler) buildQuery(input types.SearchTracesInput) (querysv
 			StartTimeMax:  maxStartTime,
 			DurationMin:   durationMin,
 			DurationMax:   durationMax,
-			SearchDepth:   searchDepth,
+			SearchDepth:   uint32(searchDepth),
 		},
 		RawTraces: false,
 	}, nil
